# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/vendor/menu-metadata:
    get:
      operationId: VendorsController_getMenuMetadata
      parameters: []
      responses:
        '200':
          description: Vendor menu categories and active unit types.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuMetadataEnvelopeDto'
        '401': *a14
        '403': *a15
      security: *a16
      tags: *a17
  /v1/vendor/menu-categories:
    post:
      operationId: VendorsController_createMenuCategory
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuCategoryDto'
      responses:
        '201':
          description: Created (or updated on slug conflict) vendor-owned menu category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuCategoryEnvelopeDto'
        '400':
          description: Invalid menu category input.
        '401': *a14
        '403': *a15
      security: *a16
      tags: *a17
...
  /v1/vendor/menu-items:
    get:
      operationId: VendorsController_listMenuItems
      parameters: []
      responses:
        '200':
          description: Vendor-owned menu items including inactive historical items.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorMenuItemListEnvelopeDto'
        '401': *a14
        '403': *a15
      security: *a16
      tags: *a17
    post:
      operationId: VendorsController_createMenuItem
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuItemDto'
      responses:
        '201':
          description: Created vendor-owned menu item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorMenuItemEnvelopeDto'
        '400':
          description: Invalid menu item input.
        '401': *a14
        '403': *a15
      security: *a16
      tags: *a17
  /v1/vendor/menu-items/{itemId}:
    get:
      operationId: VendorsController_getMenuItem
      parameters:
        - name: itemId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Vendor-owned menu item detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorMenuItemEnvelopeDto'
        '401': *a14
        '403': *a15
        '404':
          description: Menu item was not found for this vendor.
      security: *a16
      tags: *a17
    patch:
      operationId: VendorsController_updateMenuItem
      parameters:
        - name: itemId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMenuItemDto'
      responses:
        '200':
          description: Updated vendor-owned menu item safe fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorMenuItemEnvelopeDto'
        '400':
          description: Invalid menu item input.
        '401': *a14
        '403': *a15
        '404':
          description: Menu item was not found for this vendor.
      security: *a16
      tags: *a17
  /v1/vendor/menu-items/{itemId}/activate:
    post:
      operationId: VendorsController_activateMenuItem
      parameters:
        - name: itemId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Activated a vendor-owned menu item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorMenuItemEnvelopeDto'
        '401': *a14
        '403': *a15
      security: *a16
      tags: *a17
  /v1/vendor/menu-items/{itemId}/deactivate:
    post:
      operationId: VendorsController_deactivateMenuItem
      parameters:
        - name: itemId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Deactivated a vendor-owned menu item without deleting history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorMenuItemEnvelopeDto'
        '401': *a14
        '403': *a15
      security: *a16
      tags: *a17
...
  /v1/admin/unit-types:
    get:
      operationId: AdminUnitTypesController_list
      parameters: []
      responses:
        '200':
          description: All unit types including inactive ones.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitTypeListEnvelopeDto'
        '401': &a22
          description: Missing, invalid, or expired Supabase JWT.
        '403': &a23
          description: Admin role is required.
      security: &a24
        - supabaseAuth: []
      tags: &a25
        - admin
    post:
      operationId: AdminUnitTypesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUnitTypeDto'
      responses:
        '201':
          description: Created unit type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitTypeEnvelopeDto'
        '400':
          description: Invalid unit type input or duplicate code.
        '401': *a22
        '403': *a23
      security: *a24
      tags: *a25
  /v1/admin/unit-types/{id}:
    patch:
      operationId: AdminUnitTypesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUnitTypeDto'
      responses:
        '200':
          description: Updated unit type (code is immutable).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitTypeEnvelopeDto'
        '400':
          description: Invalid unit type input.
        '401': *a22
        '403': *a23
        '404':
          description: Unit type was not found.
      security: *a24
      tags: *a25
  /v1/admin/vendors/{vendorId}/menu-metadata:
    get:
      operationId: AdminVendorMenuController_getMenuMetadata
      parameters:
        - name: vendorId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Vendor menu categories and active unit types.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuMetadataEnvelopeDto'
        '401': &a26
          description: Missing, invalid, or expired Supabase JWT.
        '403': &a27
          description: Admin role is required.
        '404':
          description: Vendor was not found.
      security: &a28
        - supabaseAuth: []
      tags: &a29
        - admin
  /v1/admin/vendors/{vendorId}/menu-categories:
    post:
      operationId: AdminVendorMenuController_createMenuCategory
      parameters:
        - name: vendorId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuCategoryDto'
      responses:
        '201':
          description: Created (or updated on slug conflict) vendor menu category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuCategoryEnvelopeDto'
        '400':
          description: Invalid menu category input.
        '401': *a26
        '403': *a27
        '404':
          description: Vendor was not found.
      security: *a28
      tags: *a29
  /v1/admin/vendors/{vendorId}/menu-items:
    get:
      operationId: AdminVendorMenuController_listMenuItems
      parameters:
        - name: vendorId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Vendor menu items including inactive historical items.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorMenuItemListEnvelopeDto'
        '401': *a26
        '403': *a27
        '404':
          description: Vendor was not found.
      security: *a28
      tags: *a29
    post:
      operationId: AdminVendorMenuController_createMenuItem
      parameters:
        - name: vendorId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuItemDto'
      responses:
        '201':
          description: Created vendor menu item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorMenuItemEnvelopeDto'
        '400':
          description: Invalid menu item input.
        '401': *a26
        '403': *a27
        '404':
          description: Vendor was not found.
      security: *a28
      tags: *a29
  /v1/admin/vendors/{vendorId}/menu-items/{itemId}:
    get:
      operationId: AdminVendorMenuController_getMenuItem
      parameters:
        - name: itemId
          required: true
          in: path
          schema:
            format: uuid
            type: string
        - name: vendorId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Vendor menu item detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorMenuItemEnvelopeDto'
        '401': *a26
        '403': *a27
        '404':
          description: Menu item was not found for this vendor.
      security: *a28
      tags: *a29
    patch:
      operationId: AdminVendorMenuController_updateMenuItem
      parameters:
        - name: itemId
          required: true
          in: path
          schema:
            format: uuid
            type: string
        - name: vendorId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMenuItemDto'
      responses:
        '200':
          description: Updated vendor menu item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorMenuItemEnvelopeDto'
        '400':
          description: Invalid menu item input.
        '401': *a26
        '403': *a27
        '404':
          description: Menu item was not found for this vendor.
      security: *a28
      tags: *a29
  /v1/admin/vendors/{vendorId}/menu-items/{itemId}/activate:
    post:
      operationId: AdminVendorMenuController_activateMenuItem
      parameters:
        - name: itemId
          required: true
          in: path
          schema:
            format: uuid
            type: string
        - name: vendorId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Activated a vendor menu item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorMenuItemEnvelopeDto'
        '401': *a26
        '403': *a27
        '404':
          description: Menu item was not found for this vendor.
      security: *a28
      tags: *a29
  /v1/admin/vendors/{vendorId}/menu-items/{itemId}/deactivate:
    post:
      operationId: AdminVendorMenuController_deactivateMenuItem
      parameters:
        - name: itemId
          required: true
          in: path
          schema:
            format: uuid
            type: string
        - name: vendorId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Deactivated a vendor menu item without deleting history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorMenuItemEnvelopeDto'
        '401': *a26
        '403': *a27
        '404':
          description: Menu item was not found for this vendor.
      security: *a28
      tags: *a29
...
  /v1/catalog/vendors/{vendorId}:
    get:
      operationId: CatalogController_getVendor
      parameters: []
      responses:
        '200':
          description: Approved active vendor detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogVendorEnvelopeDto'
        '400':
          description: Invalid vendor ID.
        '404':
          description: Vendor not found or unavailable.
      tags: *a30
...
  /v1/admin/vendors/{vendorId}:
    get:
      operationId: AdminController_getVendor
      parameters: []
      responses:
        '200':
          description: ''
        '401': *a80
        '403': *a81
      security: *a82
      tags: *a83
    patch:
      operationId: AdminController_updateVendor
      parameters: []
      responses:
        '200':
          description: ''
        '401': *a80
        '403': *a81
      security: *a82
      tags: *a83
...
info:
  title: Meal Direct API
  description: Stable REST API for Meal Direct web and future mobile clients.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    supabaseAuth:
      scheme: bearer
      bearerFormat: Supabase JWT
      type: http
      description: Supabase Auth access token.
    idempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key
      description: Required for important mutations such as order creation and payments.
    requestId:
      type: apiKey
      in: header
      name: x-request-id
      description: Optional caller-provided request ID. The API echoes or generates one.
    operationsToken:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Internal operations token. Temporary until Super Admin JWT/RBAC is
        implemented.
  schemas:
...
    MenuCategoryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vendorId:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        active:
          type: boolean
        displayOrder:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - vendorId
        - name
        - slug
        - active
        - displayOrder
        - createdAt
        - updatedAt
    UnitTypeDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        displayName:
          type: string
        countsTowardSpoonLimit:
          type: boolean
        triggersTakeawayFee:
          type: boolean
        maxQuantity:
          type: number
          nullable: true
        active:
          type: boolean
      required:
        - id
        - code
        - displayName
        - countsTowardSpoonLimit
        - triggersTakeawayFee
        - active
    MenuMetadataDto:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: '#/components/schemas/MenuCategoryDto'
        unitTypes:
          type: array
          items:
            $ref: '#/components/schemas/UnitTypeDto'
      required:
        - categories
        - unitTypes
    MenuMetadataEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/MenuMetadataDto'
      required:
        - data
    CreateMenuCategoryDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
        displayOrder:
          type: number
          minimum: 0
      required:
        - name
    MenuCategoryEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/MenuCategoryDto'
      required:
        - data
...
    VendorMenuItemDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vendorId:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
          nullable: true
        categoryName:
          type: string
          nullable: true
        unitTypeId:
          type: string
          format: uuid
        unitCode:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        priceKobo:
          type: number
        countsTowardSpoonLimit:
          type: boolean
          description: When true, this item contributes to takeaway packaging rules and
            fees.
        requiresSoup:
          type: boolean
          description: When true, ordering this item requires picking one of the vendor's
            soups.
        active:
          type: boolean
        displayOrder:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - vendorId
        - unitTypeId
        - unitCode
        - name
        - priceKobo
        - countsTowardSpoonLimit
        - requiresSoup
        - active
        - displayOrder
        - createdAt
        - updatedAt
    VendorMenuItemListEnvelopeDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/VendorMenuItemDto'
      required:
        - data
    CreateMenuItemDto:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
          nullable: true
        unitTypeId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 180
          minLength: 1
        description:
          type: string
          maxLength: 1000
          nullable: true
        imageUrl:
          type: string
          maxLength: 2048
          nullable: true
        priceKobo:
          type: number
          minimum: 0
        requiresSoup:
          type: boolean
          default: false
          description: When true, ordering this item requires the customer to pick one of
            the vendor's active soups.
        displayOrder:
          type: number
          minimum: 0
      required:
        - unitTypeId
        - name
        - priceKobo
    VendorMenuItemEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/VendorMenuItemDto'
      required:
        - data
    UpdateMenuItemDto:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
          nullable: true
        unitTypeId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 180
          minLength: 1
        description:
          type: string
          maxLength: 1000
          nullable: true
        imageUrl:
          type: string
          maxLength: 2048
          nullable: true
        priceKobo:
          type: number
          minimum: 0
        requiresSoup:
          type: boolean
        displayOrder:
          type: number
          minimum: 0
...
    UnitTypeListEnvelopeDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UnitTypeDto'
      required:
        - data
    CreateUnitTypeDto:
      type: object
      properties:
        code:
          type: string
          maxLength: 40
          minLength: 1
          pattern: ^[a-z0-9_]+$
          description: Immutable lowercase code, e.g. spoon, piece, plate.
        displayName:
          type: string
          maxLength: 120
          minLength: 1
        countsTowardSpoonLimit:
          type: boolean
          default: false
          description: When true, quantities count toward the three-spoon takeaway package
            limit.
        triggersTakeawayFee:
          type: boolean
          default: false
          description: When true, an order containing this item pulls the flat
            takeaway/service fee, independent of the spoon-limit cap.
        maxQuantity:
          type: number
          minimum: 1
          nullable: true
          description: Maximum quantity per order line for this unit type; null means
            unlimited.
      required:
        - code
        - displayName
    UnitTypeEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/UnitTypeDto'
      required:
        - data
    UpdateUnitTypeDto:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 120
          minLength: 1
        countsTowardSpoonLimit:
          type: boolean
        triggersTakeawayFee:
          type: boolean
        maxQuantity:
          type: number
          minimum: 1
          nullable: true
        active:
          type: boolean
    CatalogSoupOptionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      required:
        - id
        - name
    CatalogVendorDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        campusId:
          type: string
          format: uuid
        displayName:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        kitchenLocation:
          type: string
          nullable: true
        defaultDeliveryMode:
          type: string
          enum:
            - meal_direct_rider
            - vendor_delivery
        soupOptions:
          description: The vendor's active soups; rendered as the picker for items
            requiring a soup.
          type: array
          items:
            $ref: '#/components/schemas/CatalogSoupOptionDto'
      required:
        - id
        - campusId
        - displayName
        - slug
        - defaultDeliveryMode
        - soupOptions
...
    CatalogVendorEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/CatalogVendorDto'
      required:
        - data
...